+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Sanitize XML-bulk records before saving by removing invalid fields. Fixes UIHAADM-47.
 * Move Storage engines/Transformation pipelines/Transformation steps into settings (although they are not yet functional). Fixes UIHAADM-43.
 * Remove `currentStatus` from submitted records: this is both read-only (so ignored when correct) and controlled (so it causes problems when it is not). Fixes UIHAADM-49.
+* Prevent UI from attempting to sort Old Jobs on `seconds` column. Fixes UIHAADM-51.
 
 ## [0.1.0](https://github.com/folio-org/ui-harvester-admin/tree/v0.1.0) (2022-07-19)